Version 4: soporte para coeficientes decimales usando double

+++ EcuacionSegundoGradoMain/src/ecuacionsegundogradomain/EcuacionSegundoGradoMain.java
@@ -7,17 +7,17 @@ public class EcuacionSegundoGradoMain {
     public static void main(String[] args) {
         
         Scanner sc = new Scanner(System.in);
-        int a, b, c;
+        double a, b, c;
         double x1, x2, discriminante;
         
         System.out.print("Ingrese a: ");
-        a = sc.nextInt();
+        a = sc.nextDouble();
         
         System.out.print("Ingrese b: ");
-        b = sc.nextInt();
+        b = sc.nextDouble();
         
         System.out.print("Ingrese c: ");
-        c = sc.nextInt();
+        c = sc.nextDouble();
         
         if (a == 0) {
             System.out.println("No es una ecuacion de segundo grado porque a no puede ser 0.");
